Infinite Int

+++ scheme/include/object.h
@@ -49,7 +49,7 @@ object make_object( uint type );
 object make_nil( void );
 object make_empty_list(void);
 object make_boolean(void);
-object make_integer(int n);
+object make_integer(long n);
 object make_symbol(string sym);
 object make_char(char ch);
 object make_pair(pair pa);
